test: add Github actions tests, fix non-compiling code
This change adds Github actions that'll run
on every pull request, push to main. This will help
curb the case in which we have non-compiling code or
code that fails to pass tests.

Also while here fixed the non-compiling code with the intention
of a replacement that wasn't fully implemented in PR #3

Updates #5